requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.24.5  
langchain-text-splitters==0.2.1 
redis==5.0.7
httpx==0.27.0
sentence-transformers==2.7.0
transformers==4.41.2
google-generativeai==0.7.1
numpy==1.26.4
rank-bm25==0.2.2
scikit-learn==1.5.0